writing tests

+++ src/main/java/dev/zu/bookreview/controller/BookController.java
@@ -2,26 +2,32 @@ package dev.zu.bookreview.controller;
 
 import dev.zu.bookreview.model.Book;
 import dev.zu.bookreview.service.BookService;
-import dev.zu.bookreview.service.ReviewService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
+import java.net.http.HttpResponse;
 import java.util.List;
 
 @RestController
 @RequestMapping("/books")
 public class BookController {
+
+    private final BookService bookService;
+
     @Autowired
-    private BookService bookService;
-    @Autowired
-    private ReviewService reviewService;
+    public BookController(BookService bookService) {
+        this.bookService = bookService;
+    }
 
     @GetMapping
     public List<Book> getAllBooks() {
         return bookService.getAllBooks();
     }
 
-    @GetMapping("/{id}")
+    @GetMapping("/id/{id}")
     public Book getBookById(@PathVariable Long id) {
         return bookService.getBookById(id);
     }
@@ -37,7 +43,14 @@ public class BookController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteBook(@PathVariable  Long id) {
+    public ResponseEntity<String> deleteBook(@PathVariable  Long id) {
+        if (bookService.getBookById(id).equals(false)) {
+            return ResponseEntity
+                    .status(HttpStatus.NOT_FOUND)
+                    .body("Book wasn't able to be deletes.");
+        }
         bookService.deleteBook(id);
+        // Return message "Book deleted successfully"
+        return ResponseEntity.ok("Book deleted successfully");
     }
 }
