+++ src/main/java/dev/zu/bookreview/controller/ReviewController.java
@@ -1,12 +1,14 @@
 package dev.zu.bookreview.controller;
 
 import dev.zu.bookreview.model.Review;
-import dev.zu.bookreview.service.BookService;
 import dev.zu.bookreview.service.ReviewService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
+import java.util.UUID;
 
 @RestController
 @RequestMapping("/reviews")
@@ -24,12 +26,13 @@ public class ReviewController {
         return reviewService.getAllReviews();
     }
 
-    @GetMapping("/{id}")
+    @GetMapping("/id/{id}")
     public Review getReviewById(@PathVariable Long id) {
         if (reviewService.getReviewById(id).equals(null)) {
             // Throw Exception
             throw new IllegalArgumentException("Review not found");
         }
+
         return reviewService.getReviewById(id);
     }
 
@@ -44,7 +47,13 @@ public class ReviewController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteReviewById(@PathVariable Long id) {
-        reviewService.deleteReviewById(id);
+    public ResponseEntity<String> deleteReviewById(@PathVariable Long id) {
+            if (reviewService.getReviewById(id).equals(null)) {
+                    return ResponseEntity
+                            .status(HttpStatus.NOT_FOUND)
+                            .body("Review wasn't able to be deletes.");
+            }
+            reviewService.deleteReviewById(id);
+            return ResponseEntity.ok("Review deleted successfully");
     }
 }
